Move password logic to its own module in service
