+++ components/LoginRegister.tsx
@@ -19,61 +19,7 @@ import { useToggle } from "@mantine/hooks";
 import { useRouter } from "next/router";
 import axios, { AxiosError } from "axios";
 import Requirement from "./Requirement";
-
-const requirements = [
-  {
-    validate: (password: string) => password.length >= 8,
-    label: "Has at least 8 characters",
-  },
-  {
-    validate: (password: string) => /[0-9]/.test(password),
-    label: "Includes number",
-  },
-  {
-    validate: (password: string) => /[a-z]/.test(password),
-    label: "Includes lowercase letter",
-  },
-  {
-    validate: (password: string) => /[A-Z]/.test(password),
-    label: "Includes uppercase letter",
-  },
-  {
-    validate: (password: string) => /[\W_]/.test(password),
-    label: "Includes special symbol",
-  },
-];
-
-export const getPasswordStrength = (password: string) => {
-  let strengthAccumulator = 0;
-
-  requirements.forEach((requirement) => {
-    if (requirement.validate(password)) strengthAccumulator++;
-  });
-
-  return strengthAccumulator * (100 / requirements.length);
-};
-
-export const getStrengthColorAndPhrase = (strength: number) => {
-  const colors = ["red", "orange", "yellow", "blue", "green"];
-  const phrases = [
-    "Sucks",
-    "My grandma can hack this",
-    "Still not close yet",
-    "Fair enough",
-    "Bullet proof",
-  ];
-
-  const colorAndPhraseByPercentage = requirements.map((undefined, index) => ({
-    percentage: (index + 1) * (100 / requirements.length),
-    color: colors[index],
-    phrase: phrases[index],
-  }));
-  const strengthColorAndPhrase = colorAndPhraseByPercentage.filter(
-    (e) => e.percentage >= strength
-  )[0];
-
-  return strengthColorAndPhrase;
-};
+import { getPasswordStrength, getStrengthColorAndPhrase, passwordValidation, requirements } from "../service/password";
 
 const LoginRegister = ({
   initialType,
@@ -124,13 +70,7 @@ const LoginRegister = ({
       password: (value) => {
         if (type !== "register") return null;
 
-        if (
-          getPasswordStrength(value) <
-          (requirements.length - 1) * (100 / requirements.length)
-        )
-          return "Password must follow at least 4 of the 5 contiguous guidelines";
-
-        return null;
+        return passwordValidation(value);
       },
     },
   });
